requirements.txt
# Multi-Agent Code Interpreter Requirements
# Generated for Python 3.10+

# Core Framework
fastapi==0.122.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Google Gemini AI
google-generativeai==0.8.5

# Data Processing
pandas==2.3.3
numpy==2.2.6

# Visualization
matplotlib==3.10.7
seaborn==0.13.2

# Utilities
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.1

# Additional
aiofiles==25.1.0

